pyproject.toml
[project]
name = "sundials4py-generate"
version = "0.1.0"
description = "Python code to generate some of the sundials4py bindings to the SUNDIALS suite of nonlinear and differential/algebraic equation solvers."
authors = [
    { name = "SUNDIALS Developers", email = "sundials-users@llnl.gov" }
]
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Mathematics"
]
dependencies = [
    "litgen@git+https://github.com/sundials-codes/litgen.git",
    "pyyaml",
]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
]

[project.urls]
Homepage = "https://computing.llnl.gov/projects/sundials"
Documentation = "https://sundials.readthedocs.io/"
Source = "https://github.com/LLNL/sundials"

[tool.black]
line-length = 99
skip_magic_trailing_comma = true
